Move length targeting test to nocover
